refactor hidden atribute, add frontend crud, improve context menu

+++ backend/src/app/category/category.service.ts
@@ -1,10 +1,12 @@
-import { Injectable } from '@nestjs/common';
 import { CreateCategoryDto } from '@category/dto/create-category.dto';
+import { GetAllQuery } from '@category/dto/get-all.query';
+import { RemoveQuery } from '@category/dto/remove.query';
 import { UpdateCategoryDto } from '@category/dto/update-category.dto';
-import { InjectRepository } from '@nestjs/typeorm';
 import { CategoryEntity } from '@category/entity/category.entity';
-import { Repository } from 'typeorm';
 import { ErrorService } from '@error/error.service';
+import { Injectable } from '@nestjs/common';
+import { InjectRepository } from '@nestjs/typeorm';
+import { FindManyOptions, FindOptionsWhere, IsNull, Not, Repository } from 'typeorm';
 
 @Injectable()
 export class CategoryService {
@@ -14,9 +16,19 @@ export class CategoryService {
     private readonly errorService: ErrorService,
   ) {}
 
-  async all() {
+  private readonly UNIQ_KEYS = ['article', 'code'];
+  private readonly ERROR_TRANSLATES = {
+    article: (value: string | number) => `артикулом - ${value}`,
+    code: (value: string | number) => `кодом - ${value}`,
+  };
+
+  async all(query: GetAllQuery) {
     try {
-      const categories = await this.categoryRepository.find();
+      const whereExpression: FindManyOptions<CategoryEntity> = query.onlyHidden
+        ? { where: { dateDeleted: Not(IsNull()) }, withDeleted: true }
+        : { withDeleted: false };
+
+      const categories = await this.categoryRepository.find(whereExpression);
 
       return this.errorService.success('Категории успешно получены', { categories });
     } catch (e) {
@@ -37,6 +49,13 @@ export class CategoryService {
   async create(dto: CreateCategoryDto) {
     try {
       const categoryEntity = this.categoryRepository.create(dto);
+
+      await this.checkDuplicateAndThrow(categoryEntity);
+
+      if (dto.hidden) {
+        categoryEntity.dateDeleted = new Date();
+      }
+
       const category = await this.categoryRepository.save(categoryEntity);
 
       return this.errorService.success('Категория успешно создана', { category });
@@ -47,7 +66,17 @@ export class CategoryService {
 
   async update(id: string, dto: UpdateCategoryDto) {
     try {
-      const category = await this.categoryRepository.update({ id }, dto);
+      const categoryEntity = this.categoryRepository.create({ ...dto, id });
+
+      await this.checkDuplicateAndThrow(categoryEntity);
+
+      if (dto.hidden) {
+        categoryEntity.dateDeleted = new Date();
+      } else {
+        categoryEntity.dateDeleted = null;
+      }
+
+      const category = await this.categoryRepository.update({ id }, categoryEntity);
 
       return this.errorService.success('Категория успешно обновлена', { category });
     } catch (e) {
@@ -55,13 +84,42 @@ export class CategoryService {
     }
   }
 
-  async remove(id: string) {
+  async remove(id: string, query: RemoveQuery) {
     try {
-      const category = await this.categoryRepository.delete({ id });
+      if (query.hard) {
+        const category = await this.categoryRepository.delete({ id });
+        return this.errorService.success('Категория успешно удалена', { category });
+      }
 
-      return this.errorService.success('Категория успешно удалена', { category });
+      const category = await this.categoryRepository.update({ id }, { dateDeleted: new Date() });
+      return this.errorService.success('Категория успешно скрыта', { category });
     } catch (e) {
       throw this.errorService.internal('Ошибка удаления категории', e.message);
+    }
+  }
+
+  private async checkDuplicateAndThrow(dto: CategoryEntity) {
+    for (const uniqKey of this.UNIQ_KEYS) {
+      if (!dto[uniqKey]) {
+        continue;
+      }
+
+      const expression: FindOptionsWhere<CategoryEntity> = { [uniqKey]: dto[uniqKey] };
+
+      if (dto.id) {
+        expression.id = Not(dto.id);
+      }
+
+      const categoryExists = await this.categoryRepository.findOne({
+        where: expression,
+        withDeleted: true,
+      });
+
+      if (categoryExists) {
+        const translate: string = this.ERROR_TRANSLATES[uniqKey](dto[uniqKey]);
+
+        throw new Error(`Категория с ${translate} уже существует`);
+      }
     }
   }
 }
